+++ apps/chunithm-scraper/src/main.ts
@@ -2,30 +2,32 @@ import fs from "node:fs/promises";
 
 import { Browser } from "playwright";
 
-import { ImgGenInput } from "@repo/types/chuni";
+import { FullPlayDataInput, ImgGenInput } from "@repo/types/chuni";
 
 import type { RawImageGen } from "@/app/chuni-web/src/lib/types.js";
 
 import { db } from "./db.js";
 import { environment } from "./environment.js";
 import { readHiddenCharts } from "./hidden.js";
-import { logger } from "./logger.js";
 import { recordToGenInput } from "./parser/music.js";
-import { PwPage } from "./playwright.js";
+import { handlePwError, Runner } from "./runner.js";
+import { fetchPath } from "./scraper.js";
 import { login } from "./steps/1-login.js";
 import { scrapePlayerData } from "./steps/2-playerdata.js";
 import { scrapeMusicRecord } from "./steps/3-music.js";
 import { saveToDatabase } from "./steps/6-savedb.js";
 import { generateImage } from "./steps/7-image.js";
 import { sendDiscordImage } from "./steps/8-discord.js";
+import { sendFiles } from "./utils/discord.js";
 import { downloadImageAsBase64 } from "./utils/image.js";
+import { logger } from "./utils/logger.js";
 
 export async function main(jobId: number | undefined, browser: Browser) {
   // * Step 0: Preparation
   const hiddenCharts = await readHiddenCharts();
 
   const page = await browser.newPage();
-  const pwPage = new PwPage(page);
+  const runner = new Runner();
 
   // Create folder "outputs" if not exists
   try {
@@ -37,28 +39,32 @@ export async function main(jobId: number | undefined, browser: Browser) {
   }
 
   // * Step 1: Login
-  await pwPage.runStep("Step 1: Login", (page) => login(page));
-  await pwPage.runStep("Prepare for Step 2", async (page) => {
-    await page.getByRole("link", { name: "PLAYER DATA" }).click();
-    await page.waitForURL(
-      "https://chunithm-net-eng.com/mobile/home/playerData",
-    );
-  });
+  await runner.runStep(
+    "Step 1: Login",
+    () => login(page),
+    (ctx) => handlePwError(ctx, page),
+  );
 
   // * Step 2: Player Data
-  const { playerData, playerDataHtml } = await pwPage.runStep(
-    "Step 2: Player Data",
-    scrapePlayerData,
+  const playerDataPage = await runner.runStep(
+    "Step 2.1: Fetch Player Data",
+    () => fetchPath(page, "home/playerData"),
+  );
+
+  const { playerData, playerDataHtml } = await runner.runStep(
+    "Step 2.2: Process Player Data",
+    () => scrapePlayerData(playerDataPage),
+    (ctx) => handlePwError(ctx),
   );
 
   // * Step 3: Music Record (For Rating and All)
-  const { recordData, recordHtml } = await pwPage.runStep(
+  const { recordData, recordHtml } = await runner.runStep(
     "Step 3: Scrape Music Records",
-    scrapeMusicRecord,
+    () => scrapeMusicRecord(page),
   );
 
   // * Step 4: Create JSON for Image Generation
-  const { imgGenInput, fileName } = await pwPage.runStep(
+  const { imgGenInput, imgGenFileName } = await runner.runStep(
     "Step 4: Create JSON for Image Generation",
     async () => {
       const charaImageData = await downloadImageAsBase64(
@@ -77,19 +83,32 @@ export async function main(jobId: number | undefined, browser: Browser) {
         hidden: hiddenCharts || undefined,
       } satisfies ImgGenInput;
 
-      const fileName = `${playerData.lastPlayed.toISOString()}-${jobId}.json`;
+      const imgGenFileName = `${playerData.lastPlayed.toISOString()}-${jobId}.json`;
 
       await fs.writeFile(
-        `outputs/${fileName}`,
+        `outputs/${imgGenFileName}`,
         JSON.stringify(imgGenInput, null, 2),
       );
 
-      return { imgGenInput, fileName };
+      const fullData = {
+        ...imgGenInput,
+        allRecords: recordData.allRecords
+          .filter((r) => r.score > 0)
+          .map(recordToGenInput),
+      } satisfies FullPlayDataInput;
+
+      const fullDataFileName = `full-${imgGenFileName}`;
+      await fs.writeFile(
+        `outputs/${fullDataFileName}`,
+        JSON.stringify(fullData, null, 2),
+      );
+
+      return { imgGenInput, imgGenFileName, fullData, fullDataFileName };
     },
   );
 
   // * Step 5: Calculate Rating
-  const rawImgGen = await pwPage.runStep(
+  const rawImgGen = await runner.runStep(
     "Step 5: Calculate Rating",
     async () => {
       if (!environment.IMAGE_GEN_URL) {
@@ -117,11 +136,13 @@ export async function main(jobId: number | undefined, browser: Browser) {
 
       return (await res.json()) as RawImageGen;
     },
+    handlePwError,
+    3,
   );
 
   // * Step 6: Save data to DB
   if (db) {
-    await pwPage
+    await runner
       .runStep("Step 6: Save to DB", async () => {
         await saveToDatabase(jobId!, db!, {
           playerData,
@@ -134,19 +155,26 @@ export async function main(jobId: number | undefined, browser: Browser) {
       })
       .catch((err) => {
         logger.error(`Step 6 Error: ${err}`);
+        sendFiles("Step 6 Error: Failed saving to database", [
+          {
+            blob: new Blob([`${err}`], { type: "text/plain" }),
+            fileName: `error-step6-${Date.now()}.txt`,
+          },
+        ]);
       });
   } else {
     logger.warn("Database Mode disabled, skipped saving to DB");
   }
 
   // * Step 7: Generate Image
-  const outputLocation = await pwPage.runStep(
+  const outputLocation = await runner.runStep(
     "Step 7: Generate Image",
-    (page) => generateImage(page, fileName),
+    () => generateImage(page, imgGenFileName),
+    (ctx) => handlePwError(ctx, page),
   );
 
   // * Step 8: Send Image to Discord
-  await pwPage.runStep("Step 8: Send Image to Discord", () =>
+  await runner.runStep("Step 8: Send Image to Discord", () =>
     sendDiscordImage(outputLocation, playerData, rawImgGen),
   );
 }