+++ apps/chunithm-scraper/src/steps/3-music.ts
@@ -1,8 +1,9 @@
 import { JSDOM } from "jsdom";
 import { Page } from "playwright";
 
-import { logger } from "../logger.js";
 import { parseMusic, parseRecord } from "../parser/music.js";
+import { fetchPath } from "../scraper.js";
+import { logger } from "../utils/logger.js";
 
 function fillMarkInfo(
   original: Array<ReturnType<typeof parseMusic>>,
@@ -29,20 +30,23 @@ function fillMarkInfo(
 }
 
 export async function scrapeMusicRecord(page: Page) {
-  await page.getByRole("link", { name: "MUSIC FOR RATING" }).click();
-  // Best Songs
-  await page.waitForURL(
-    "https://chunithm-net-eng.com/mobile/home/playerData/ratingDetailBest/",
+  const ratingBestPageHTML = await fetchPath(
+    page,
+    "home/playerData/ratingDetailBest",
   );
-  await page.waitForLoadState("networkidle");
 
   let totalHTML = "<!-- BEST -->\n";
 
-  // warning: error if no best songs
-  const best = page.locator(".box05.w400");
-  const bestHTML = await best.innerHTML();
-  totalHTML += bestHTML;
-  const bestDom = new JSDOM(bestHTML);
+  const best = new JSDOM(ratingBestPageHTML).window.document.querySelector(
+    ".box05.w400",
+  );
+  if (!best) {
+    throw new Error(
+      `Step 3 Error: Best songs not found in ratingBest page. You need at least 1 best song for this to work.\nHTML: ${ratingBestPageHTML}`,
+    );
+  }
+  totalHTML += best.innerHTML;
+  const bestDom = new JSDOM(best.innerHTML);
 
   const bestMusic = [] as Array<ReturnType<typeof parseMusic>>;
   for (const element of bestDom.window.document.querySelectorAll(
@@ -52,18 +56,23 @@ export async function scrapeMusicRecord(page: Page) {
   }
   logger.log("Step 3.1.1: Best Songs done");
 
-  await page.getByRole("link", { name: "Current" }).click();
   // Current (New) Songs
-  await page.waitForURL(
-    "https://chunithm-net-eng.com/mobile/home/playerData/ratingDetailRecent/",
+  const ratingCurrentPageHTML = await fetchPath(
+    page,
+    "home/playerData/ratingDetailRecent",
   );
-  await page.waitForLoadState("networkidle");
 
   // warning: error if no current songs
-  const current = page.locator(".box05.w400");
-  const currentHTML = await current.innerHTML();
-  totalHTML += `\n<!-- NEW -->\n${currentHTML}`;
-  const currentDom = new JSDOM(currentHTML);
+  const current = new JSDOM(
+    ratingCurrentPageHTML,
+  ).window.document.querySelector(".box05.w400");
+  if (!current) {
+    throw new Error(
+      `Step 3 Error: Current songs not found in ratingCurrent page. You need at least 1 current song for this to work.\nHTML: ${ratingCurrentPageHTML}`,
+    );
+  }
+  totalHTML += `\n<!-- NEW -->\n${current.innerHTML}`;
+  const currentDom = new JSDOM(current.innerHTML);
 
   const currentMusic = [] as Array<ReturnType<typeof parseMusic>>;
   for (const element of currentDom.window.document.querySelectorAll(
@@ -73,22 +82,23 @@ export async function scrapeMusicRecord(page: Page) {
   }
   logger.log("Step 3.1.2: Current Songs done");
 
-  await page.getByRole("link", { name: "Selection" }).click();
   // Selection
-  await page.waitForURL(
-    "https://chunithm-net-eng.com/mobile/home/playerData/ratingDetailNext/",
+  const ratingSelectionPageHTML = await fetchPath(
+    page,
+    "home/playerData/ratingDetailNext",
   );
-  await page.waitForLoadState("networkidle");
 
-  const selection = await page.locator(".w420.box01").all();
+  const selection = new JSDOM(
+    ratingSelectionPageHTML,
+  ).window.document.querySelectorAll(".w420.box01");
 
   if (selection.length !== 2) {
     throw new Error(
-      `Step 3 Error: Expected 2 elements in selection page, got ${selection.length}`,
+      `Step 3 Error: Expected 2 elements in selection page, got ${selection.length}\nPage Content: ${ratingSelectionPageHTML}`,
     );
   }
 
-  const selectionBestHTML = await selection[0].innerHTML();
+  const selectionBestHTML = selection[0].innerHTML;
   totalHTML += `\n<!-- SELECTION BEST -->\n${selectionBestHTML}`;
   const selectionBestDom = new JSDOM(selectionBestHTML);
   const selectionBestMusic = [] as Array<ReturnType<typeof parseMusic>>;
@@ -98,7 +108,7 @@ export async function scrapeMusicRecord(page: Page) {
     selectionBestMusic.push(parseMusic(element));
   }
 
-  const selectionCurrentHTML = await selection[1].innerHTML();
+  const selectionCurrentHTML = selection[1].innerHTML;
   totalHTML += `\n<!-- SELECTION BEST -->\n${selectionCurrentHTML}`;
   const selectionCurrentDom = new JSDOM(selectionCurrentHTML);
   const selectionCurrentMusic = [] as Array<ReturnType<typeof parseMusic>>;
@@ -110,38 +120,47 @@ export async function scrapeMusicRecord(page: Page) {
   logger.log("Step 3.1.3: Selection Songs done");
 
   // Part 3.2: All Records
-  const allRecords = [] as Array<ReturnType<typeof parseRecord>>;
-  await page.locator(".btn_record").click();
-  await page.waitForURL("https://chunithm-net-eng.com/mobile/record/");
-  await page.waitForLoadState("networkidle");
-
-  await page.getByRole("link", { name: "RECORD BY MUSIC" }).click();
-  await page.waitForURL(
-    "https://chunithm-net-eng.com/mobile/record/musicGenre",
+  const ratingAllPageHTML = await fetchPath(page, "record/musicGenre");
+  totalHTML += `\n<!-- All Music Home Page -->\n${ratingAllPageHTML}`;
+  const tokenInput = new JSDOM(ratingAllPageHTML).window.document.querySelector(
+    "input[name=token]",
   );
-  await page.waitForLoadState("networkidle");
+  if (!tokenInput) {
+    throw new Error(
+      `Step 3.2 Error: Failed to get token input\nHTML: ${ratingAllPageHTML}`,
+    );
+  }
 
-  const btnClasses = [
-    "btn_basic",
-    "btn_advanced",
-    "btn_expert",
-    "btn_master",
-    "btn_ultimate",
-  ];
-  for (const btnClass of btnClasses) {
-    await page.locator(`.${btnClass}`).click();
-    await page.waitForLoadState("networkidle");
-    const allCategories = await page.locator(".box05.w400").all();
+  const token = (tokenInput as HTMLInputElement).value;
+
+  const allRecords = [] as Array<ReturnType<typeof parseRecord>>;
+
+  const difficulties = ["Basic", "Advanced", "Expert", "Master", "Ultima"];
+
+  for (const diff of difficulties) {
+    const ratingDiffPageHTML = await fetchPath(
+      page,
+      `record/musicGenre/send${diff}`,
+      "POST",
+      {
+        "Content-Type": "application/x-www-form-urlencoded",
+      },
+      `genre=99&token=${token}`,
+    );
+
+    const allCategories = new JSDOM(
+      ratingDiffPageHTML,
+    ).window.document.querySelectorAll(".box05.w400");
 
     if (allCategories.length !== 7) {
       throw new Error(
-        `Step 3 Error: Expected 7 elements in all categories, got ${allCategories.length}`,
+        `Step 3 Error: Expected 7 elements in all categories, got ${allCategories.length}\nPage Content: ${ratingDiffPageHTML}`,
       );
     }
 
     for (const category of allCategories) {
-      const categoryHTML = await category.innerHTML();
-      totalHTML += `\n<!-- ${btnClass} -->\n${categoryHTML}`;
+      const categoryHTML = category.innerHTML;
+      totalHTML += `\n<!-- ${diff} -->\n${categoryHTML}`;
       const categoryDom = new JSDOM(categoryHTML);
       for (const element of categoryDom.window.document.querySelectorAll(
         ".w388.musiclist_box",
@@ -150,7 +169,7 @@ export async function scrapeMusicRecord(page: Page) {
       }
     }
 
-    logger.log(`Step 3.2: ${btnClass} done`);
+    logger.log(`Step 3.2: ${diff} done`);
   }
 
   return {