+++ apps/chunithm-scraper/src/steps/6-savedb.ts
@@ -9,8 +9,8 @@ import type { BaseChartSchema, ImgGenInput } from "@repo/types/chuni";
 
 import type { Db } from "../db.js";
 import { environment } from "../environment.js";
-import { logger } from "../logger.js";
 import { recordToGenInputWithFullChain } from "../parser/music.js";
+import { logger } from "../utils/logger.js";
 import type { scrapePlayerData } from "./2-playerdata.js";
 import type { scrapeMusicRecord } from "./3-music.js";